requirements.txt
pytube==11.0.1
requests==2.26.0
pyTelegramBotAPI==4.2.0
dropbox==11.23.0
youtube_dl==2021.6.6
translators==4.9.5
telebot==0.0.4
gtts==2.2.3
dialogflow==2.0.0
rev_ai==2.13.0
todoist-python==8.1.3
datetime==4.3
heroku3==5.1.4
SpeechRecognition==3.8.1
python-telegram-bot==13.8.1
ffmpeg==1.4



